refactor(ai-service): Step 5 — backend HTTP adapter → app/integrations/backend/
The adapter that talks to the Node service is an outbound integration,
not a peer of in-process services. It belongs in app/integrations/.

Moves:
  app/services/backend_service.py  →  app/integrations/backend/client.py

app/integrations/backend/__init__.py re-exports BackendService and
the `backend_service` singleton so external callers can keep using
`from app.integrations.backend import backend_service` without
reaching into the .client module.

Importers updated (6 imports across 6 files):
  • app/main.py, app/api/health.py
  • app/services/{ai_service, __init__, full_dataset_ml_service}.py
  • app/services/llm/telemetry_llm_orchestrator.py

Plus one cross-band fix: app/storage/cache/__init__.py held a stale
`from .backend_service import …` from when both files lived as
siblings in services/. Rewritten to the absolute new path.

Validated: backend_service singleton identity preserved via package
re-export; /health 200; app.main imports clean.
