+++ acla_ai_service/app/api/health.py
@@ -5,7 +5,7 @@ Health check endpoints
 from fastapi import APIRouter
 from datetime import datetime
 from app.api.schemas.query import HealthResponse
-from app.services.backend_service import backend_service
+from app.integrations.backend.client import backend_service
 from app.core import settings
 
 router = APIRouter(tags=["health"])
