+++ acla_ai_service/app/services/__init__.py
@@ -3,7 +3,7 @@ Services module initialization
 """
 
 from .ai_service import AIService
-from .backend_service import BackendService
+from app.integrations.backend.client import BackendService
 from .tire_grip_analysis_service import TireGripAnalysisService, tire_grip_analysis_service
 from .segment_classifier_service import SegmentClassifierService, segment_classifier
